requirements.txt
awscli==1.11.44
backports.ssl-match-hostname==3.4.0.2
botocore==1.5.7
chardet==2.3.0
click==6.6
colorama==0.3.7
cryptography==1.2.3
DateTime==4.2
decorator==4.0.6
docker-compose==1.5.2
docker-py==1.8.0
docker-pycreds==0.2.1
dockerpty==0.3.4
docopt==0.6.2
docutils==0.14
enum34==1.1.2
Flask==0.12
Flask-Script==2.0.5
funcsigs==0.4
functools32==3.2.3.post2
futures==3.1.1
idna==2.0
# iotop==0.6
ipaddress==1.0.16
itsdangerous==0.24
Jinja2==2.8.1
jmespath==0.9.3
jsonschema==2.5.1
ldap3==2.1.1
Mako==1.0.6
MarkupSafe==0.23
mock==1.3.0
MySQL-python==1.2.5
ndg-httpsclient==0.4.0
neo4j-driver==1.4.0
pbr==1.8.0
pep8==1.7.0
pexpect==4.0.1
psycopg2==2.6.2
ptyprocess==0.5
py2-ipaddress==3.4.1
pyasn1==0.1.9
pycrypto==2.6.1
pymongo==3.4.0
PyMySQL==0.7.10
pyOpenSSL==0.15.1
python-dateutil==2.6.0
python-editor==1.0.3
python-ldap==2.4.28
pytz==2016.10
PyYAML==3.11
requests==2.9.1
rsa==3.4.2
s3transfer==0.1.11
simplegeneric==0.8.1
six==1.10.0
SQLAlchemy==1.1.9
texttable==0.8.1
urllib3==1.13.1
websocket-client==0.18.0
Werkzeug==0.11.15
zope.interface==4.4.2
